requirements.txt
blinker==1.9.0
Brotli==1.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
ConfigArgParse==1.7
Flask==3.1.0
flask-cors==5.0.1
Flask-Login==0.6.3
gevent==24.11.1
geventhttpclient==2.3.3
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
locust==2.33.2
MarkupSafe==3.0.2
msgpack==1.1.0
psutil==7.0.0
python-dotenv==1.0.1
pyzmq==26.3.0
requests==2.32.3
setuptools==76.0.0
urllib3==2.3.0
Werkzeug==3.1.3
zope.event==5.0
zope.interface==7.2
